requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.8.1=pypi_0
astor=0.8.0=pypi_0
blas=1.0=mkl
ca-certificates=2019.10.16=0
cachetools=3.1.1=pypi_0
certifi=2019.9.11=py36_0
cffi=1.11.5=py36h74b6da3_1
chardet=3.0.4=pypi_0
cudatoolkit=10.1.168=0
cycler=0.10.0=py36h009560c_0
freetype=2.9.1=ha9979f8_1
gast=0.2.2=pypi_0
google-auth=1.7.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=pypi_0
grpcio=1.25.0=pypi_0
h5py=2.10.0=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
idna=2.8=pypi_0
intel-openmp=2019.4=245
jpeg=9b=hb83a4c4_2
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.0.1=py36h6538335_0
libpng=1.6.34=h79bbb47_0
libtiff=4.0.9=h36446d0_2
markdown=3.1.1=pypi_0
matplotlib=3.0.0=pypi_0
mkl=2018.0.3=1
mkl_fft=1.0.6=py36hdbbee80_0
mkl_random=1.0.1=py36h77b88f5_1
ninja=1.8.2=py36he980bc4_1
numpy=1.17.4=pypi_0
numpy-base=1.15.4=py36h8128ebf_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
opencv-python=3.4.3.18=pypi_0
openssl=1.0.2t=vc14h62dcd97_0
opt-einsum=3.1.0=pypi_0
pillow=5.2.0=py36h08bbbbd_0
pip=19.3.1=py36_0
protobuf=3.6.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycparser=2.19=py36_0
pymysql=0.9.3=pypi_0
pyparsing=2.4.5=py_0
pyqt=5.9.2=py36h6538335_2
python=3.6.7=h33f27b4_1
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.6_cuda101_cudnn7_0
pytz=2019.3=py_0
qt=5.9.6=vc14h62aca36_0
requests=2.22.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scipy=1.1.0=py36h4f6bf74_1
setuptools=41.6.0=py36_0
sip=4.19.8=py36h6538335_0
six=1.13.0=py36_0
sqlite=3.25.2=hfa6e2cd_0
tbb=2018.0.5=he980bc4_0
tbb4py=2018.0.5=py36he980bc4_0
tensorboard=2.0.1=pypi_0
tensorflow-estimator=2.0.1=pypi_0
tensorflow-gpu=2.0.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.2=py36_cu101
tornado=5.1.1=py36hfa6e2cd_0
tqdm=4.39.0=pypi_0
urllib3=1.25.7=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_0
werkzeug=0.16.0=pypi_0
wheel=0.33.6=py36_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.11.1=pypi_0
zlib=1.2.11=h8395fce_2
